+++ app/helpers/members_helper.rb
@@ -1,5 +1,6 @@
 module MembersHelper
   def create_add_links(team)
+    return ->(_){} unless can? :create, :members_list
     lambda do |mutant|
       path = team_members_path(team_id: team.to_param, id: mutant.to_param)
       button_link('+', path, 'btn-success', remote: true, method: :post)
@@ -7,6 +8,7 @@ module MembersHelper
   end
 
   def create_remove_links(team)
+    return ->(_){} unless can? :destroy, :members_list
     lambda do |mutant|
       path = team_member_path(team_id: team.to_param, id: mutant.to_param)
       button_link('-', path, 'btn-danger', remote: true, method: :delete)
